+++ libs/qsynedit/qsynedit/painter.cpp
@@ -407,7 +407,7 @@ void QSynEditPainter::paintToken(
                         }
                         if (tryLigature) {
                             QString textToPaint = glyph;
-                            while(i+1<glyphStartCharList.length()) {
+                            while(i+1<endGlyph) {
                                 int glyphStart = glyphStartCharList[i+1];
                                 int glyphLen = calcSegmentInterval(glyphStartCharList,lineText.length(),i+1);
                                 QString glyph2 = lineText.mid(glyphStart,glyphLen);
@@ -428,43 +428,36 @@ void QSynEditPainter::paintToken(
                                 mPainter->setFont(font);
                                 fontInited = true;
                             }
+                            //qDebug()<<"paint 1:"<<textToPaint;
                             mPainter->drawText(nX,rcToken.bottom()-mPainter->fontMetrics().descent() , textToPaint);
                             drawed = true;
                         }
                     }
                     if (!drawed) {
-                        if (glyph.length()==1 && glyph.front().unicode()<128) {
-                            QString ch;
+                        if (glyph.length()>0) {
+                            QString textToPaint = glyph;
                             int padding=0;
                             if (showGlyphs) {
                                 switch(glyph.front().unicode()) {
                                 case '\t':
-                                    ch=TabGlyph;
+                                    textToPaint=TabGlyph;
                                     padding=(glyphWidth-1)/2*mEdit->mCharWidth;
                                     break;
                                 case ' ':
-                                    ch=SpaceGlyph;
+                                    textToPaint=SpaceGlyph;
                                     break;
                                 default:
-                                    ch=glyph;
+                                    break;
                                 }
-                            } else {
-                                ch=glyph;
                             }
-                            if (ch!=" " && ch!="\t") {
+                            if (textToPaint!=" " && textToPaint!="\t") {
                                 if (!fontInited) {
                                     mPainter->setFont(font);
                                     fontInited = true;
                                 }
-                                mPainter->drawText(nX+padding,rcToken.bottom()-mPainter->fontMetrics().descent() , ch);
+                                //qDebug()<<"Drawing"<<textToPaint;
+                                mPainter->drawText(nX+padding,rcToken.bottom()-mPainter->fontMetrics().descent() , textToPaint);
                             }
-                            //qDebug()<<"Drawing"<<glyph<<nX<<glyphWidth;
-                        } else {
-                            if (!fontInited) {
-                                mPainter->setFont(font);
-                                fontInited = true;
-                            }
-                            mPainter->drawText(nX,rcToken.bottom()-mPainter->fontMetrics().descent() , glyph);
                         }
                         drawed = true;
                     }
@@ -534,8 +527,7 @@ void QSynEditPainter::paintEditAreas(const EditingAreaList &areaList)
 
 void QSynEditPainter::paintHighlightToken(const QString& lineText,
                                           const QList<int> &glyphStartCharList,
-                                          const QList<int> &glyphStartPositionsList,
-                                          bool bFillToEOL)
+                                          const QList<int> &glyphStartPositionsList)
 {
     bool isComplexToken;
     int nC1, nC2, nC1Sel, nC2Sel;
@@ -625,43 +617,6 @@ void QSynEditPainter::paintHighlightToken(const QString& lineText,
                         mTokenAccu.font, mTokenAccu.showSpecialGlyphs);
         }
     }
-
-    // Fill the background to the end of this line if necessary.
-    if (bFillToEOL && rcToken.left() < rcLine.right()) {
-        if (mIsSpecialLine && colSpBG.isValid())
-            colBG = colSpBG;
-        else
-            colBG = colEditorBG();
-        if (mIsComplexLine) {
-            nX1 = fixXValue(mLineSelStart);
-            nX2 = fixXValue(mLineSelEnd);
-            if (rcToken.left() < nX1) {
-                setDrawingColors(false);
-                rcToken.setRight(nX1);
-//                if (TokenAccu.Len != 0 && TokenAccu.Style != FontStyle::fsNone)
-//                    AdjustEndRect();
-                mPainter->fillRect(rcToken,mPainter->brush());
-                rcToken.setLeft(nX1);
-            }
-            if (rcToken.left() < nX2) {
-                setDrawingColors(true);
-                rcToken.setRight(nX2);
-                mPainter->fillRect(rcToken,mPainter->brush());
-                rcToken.setLeft(nX2);
-            }
-            if (rcToken.left() < rcLine.right()) {
-                setDrawingColors(false);
-                rcToken.setRight(rcLine.right());
-                mPainter->fillRect(rcToken,mPainter->brush());
-            }
-        }  else {
-            setDrawingColors(mIsLineSelected);
-            rcToken.setRight(rcLine.right());
-//            if (TokenAccu.Len != 0 && TokenAccu.Style != FontStyle::fsNone)
-//                AdjustEndRect();
-            mPainter->fillRect(rcToken,mPainter->brush());
-        }
-    }
 }
 
 // Store the token chars with the attributes in the TokenAccu
@@ -723,7 +678,7 @@ void QSynEditPainter::addHighlightToken(
         }
         // If we can't append it, then we have to paint the old token chars first.
         if (!bCanAppend)
-            paintHighlightToken(lineText, glyphStartCharList, glyphStartPositionList, false);
+            paintHighlightToken(lineText, glyphStartCharList, glyphStartPositionList);
     }
     if (bInitFont) {
         mTokenAccu.style = style;
@@ -954,6 +909,8 @@ void QSynEditPainter::paintLines()
         mIsComplexLine = false;
         mLineSelStart = 0;
         mLineSelEnd = 0;
+
+        bool selToEnd = false;;
         // Does the selection intersect the visible area?
         if (bAnySelection && (row >= mSelStart.row) && (row <= mSelEnd.row)) {
             // Default to a fully selected line. This is correct for the smLine
@@ -981,6 +938,8 @@ void QSynEditPainter::paintLines()
                     mLineSelEnd = xpos;
                     mIsComplexLine = true;
                 }
+            } else {
+                selToEnd = true;
             }
         } //endif bAnySelection
 
@@ -991,6 +950,14 @@ void QSynEditPainter::paintLines()
         rcLine.setHeight(mEdit->mTextHeight);
 
         mIsLineSelected = (!mIsComplexLine) && (mLineSelStart > 0);
+
+        if (mIsSpecialLine && colSpBG.isValid())
+            colBG = colSpBG;
+        else
+            colBG = colEditorBG();
+        setDrawingColors(selToEnd);
+        mPainter->fillRect(rcLine,mPainter->brush());
+
         rcToken = rcLine;
 
         int lineWidth;
@@ -1200,7 +1167,7 @@ void QSynEditPainter::paintLines()
             }
             // Draw anything that's left in the TokenAccu record. Fill to the end
             // of the invalid area with the correct colors.
-            paintHighlightToken(sLine, glyphStartCharList, glyphStartPositionsList, true);
+            paintHighlightToken(sLine, glyphStartCharList, glyphStartPositionsList);
 
             //Paint editingAreaBorders
             foreach (const PEditingArea& area, areaList) {